// src/api/model/accessTokenResponse.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */

export interface AccessTokenResponse {
	/** @nullable */
	tokenType?: string | null;
	accessToken: string;
	/** @pattern ^-?(?:0|[1-9]\d*)$ */
	expiresIn: number | string;
	refreshToken: string;
}