// src/api/model/loginRequest.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * CRM.Backend.Web | v1
 * OpenAPI spec version: 1.0.0
 */

export interface LoginRequest {
	email: string;
	password: string;
	/** @nullable */
	twoFactorCode?: string | null;
	/** @nullable */
	twoFactorRecoveryCode?: string | null;
}